Fixed Wid reinterpret_cast.
at least on unix Wid is a unsigned long and reinterpret cast with loss of precision
is not possible. Of course it's still possible to fail by not storing an
unsigned long in the first place.

+++ kleopatra/kleopatraapplication.cpp
@@ -292,7 +292,7 @@ int KleopatraApplication::newInstance() {
         // Check for Parent Window id
         WId parentId = 0;
         if ( args->isSet( "parent-windowid" ) ) {
-            parentId = reinterpret_cast<WId>(args->getOption( "parent-windowid" ).toUInt());
+            parentId = reinterpret_cast<WId>(args->getOption( "parent-windowid" ).toULong());
         }
 
         // Search for local keys
